Improve Access Point management page - add all missing validations

+++ delis-web/src/main/java/dk/erst/delis/web/accesspoint/AccessPointController.java
@@ -1,8 +1,7 @@
 package dk.erst.delis.web.accesspoint;
 
 import java.security.cert.CertificateException;
-
-import javax.validation.Valid;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -11,8 +10,11 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import dk.erst.delis.data.entities.access.AccessPoint;
+import dk.erst.delis.data.entities.organisation.Organisation;
+import dk.erst.delis.task.organisation.setup.ValidationResultData;
 
 @Controller
 @RequestMapping("/accesspoint")
@@ -32,8 +34,14 @@ public class AccessPointController {
     }
 
     @PostMapping("save")
-    public String createNew(@Valid AccessPointData accessPoint, Model model) throws Exception {
-
+    public String createNew(AccessPointData accessPoint, Model model) throws Exception {
+    	ValidationResultData validationResultData = service.validate(accessPoint);
+		if (!validationResultData.isAllValid()) {
+			model.addAttribute("errorMessage", "Some fields are not valid");
+			model.addAttribute("accessPoint", accessPoint);
+			model.addAttribute("validation", validationResultData);
+			return "accesspoint/edit";
+		}
         try {
             service.saveAccessPoint(accessPoint);
         } catch (CertificateException e) {
@@ -44,19 +52,12 @@ public class AccessPointController {
         return "redirect:/accesspoint/list";
     }
 
-    @GetMapping("create/{id}")
-    public String createNew(@PathVariable(required = false) Long id, Model model) {
-        if (id == null || id == 0) {
-            model.addAttribute("accessPoint", new AccessPointData());
-        } else {
-            AccessPoint accessPoint = service.findById(id);
-            AccessPointData accessPointData = new AccessPointData();
-            service.copyToDTOEdit(accessPoint, accessPointData);
-            model.addAttribute("accessPoint", accessPointData);
-        }
+    @GetMapping("create")
+    public String createNew(Model model) {
+        model.addAttribute("accessPoint", new AccessPointData());
         return "accesspoint/edit";
     }
-
+    
     @GetMapping("update/{id}")
     public String updateAccessPoint(@PathVariable long id, Model model) {
         AccessPoint accessPoint = service.findById(id);
@@ -65,11 +66,23 @@ public class AccessPointController {
         model.addAttribute("accessPoint", accessPointData);
         return "accesspoint/edit";
     }
-
-    @GetMapping("delete/{id}")
-    public String deleteUser(@PathVariable long id, Model model) {
-        service.deleteAccessPoint(id);
-        model.addAttribute("accessPointList", service.getAccessPoints());
-        return "accesspoint/list";
-    }
+    
+	@GetMapping("delete/{id}")
+	public String deleteAccessPoint(@PathVariable long id, RedirectAttributes ra) {
+		List<Organisation> referenceList = service.getOrganisationReferencedAccessPointList(id);
+		if (!referenceList.isEmpty()) {
+			final StringBuilder sb = new StringBuilder();
+			referenceList.forEach(o -> {
+				if (sb.length() > 0) {
+					sb.append(", ");
+				}
+				sb.append(o.getName());
+			});
+			ra.addFlashAttribute("errorMessage", "Access point is used in configuration setup of " + referenceList.size() + " organisation(s). Before deletion, please switch from this access point to another. Organisations: " + sb.toString());
+		} else {
+			service.deleteAccessPoint(id);
+			ra.addFlashAttribute("message", "Access point is deleted");
+		}
+		return "redirect:/accesspoint/list";
+	}
 }
